refactor(weather): Move weather repositories to feature-specific directory
- Relocated all weather-related repository classes from lib/repositories/
to lib/features/weather/repositories/ to improve feature encapsulation.
- Updated internal and external import paths accordingly.
